Changed dest
Changed the destination to the nearby 4 locations of the pickup station rather than the exact co-ordinates of the pickup station. This would help in the case where the robots have to go to waiting area after dropping the object. 
If the dest is blocked or if any of the of the on-going path is blocked, a round-about path is being chosed